Add files via upload
Introduction:
In this project we are to use the data visualization libraries that are in python, r or java script to create interactive visualizations.  These interactive visualizations were created based on the text analysis that was conducted in python or R. Insightful conversation with ChatGPT, we engaged in a dialogue that delves into the complexities of automation's impact on society and the economy. The conversation is structured around a pivotal topic, the Economic Implications of Automation. We navigate through a series of thought-provoking questions and insights, each shedding light on distinct parts of this transformative concept. My exploration begins by unraveling the comprehensive dimensions of automation's influence on the economy. I seek to understand how automation is reshaping economic dynamics. With ChatGPT the conversation got into the ways automation alters employment patterns, job prospects, and the skills demanded by industries. It led to strategies that governments can adopt to mitigate potential job losses resulting from automation. I then asked a question on how educational and training programs can adapt to prepare individuals and the workforce for the changing job landscape. In essence, I will leverage this conversation with ChatGPT to conduct text analysis and create interactive visualizations using Python.